+++ examples/22-case-studies.md
