+++ lib/exec-suite.js
@@ -29,6 +29,7 @@ const originalAcquireDeps = require('./acquire-deps-original');
 const makeAcquireDepsFillIn = require('./acquire-deps-fill-in');
 const TestSuite = require('./TestSuite');
 const fatalRequestReply = require('./helpers/fatal-request-reply');
+const handleInjections = require('./handle-injections');
 
 ////////////////////////////////////////////////////////////////////
 
@@ -143,7 +144,8 @@ module.exports = {
         globalHooks.apply(suite, [suite]);
       }
       catch (err) {
-        console.error('\n' + colors.magenta(' => Suman warning => Could not find the "suman.hooks.js" file in your <sumanHelpersDir>.\n' +
+        console.error('\n' + colors.magenta(' => Suman warning => Could not find the "suman.hooks.js" ' +
+            'file in your <sumanHelpersDir>.\n' +
             'Create the file to remove the warning.'), '\n\n');
         if (global.sumanOpts.verbose) {
           console.error('\n', err.stack, '\n');
@@ -162,7 +164,7 @@ module.exports = {
 
         d.once('error', function (err) {
 
-          console.error(' ZOOOM ',err.stack || err);
+          console.error(' ZOOOM ', err.stack || err);
           global._writeTestError(err.stack);
 
           d.exit();
@@ -296,47 +298,19 @@ module.exports = {
                 cb.apply(suite, deps);
                 suite.__proto__.isSetupComplete = true;
 
-                debugger;
+                handleInjections(suite, function (err) {
 
-                process.nextTick(function(){
-
-                  async.eachSeries(suite.getInjections(), function (inj, cb) {
-
-                    console.log('IN SERIES (exec suite)');
-
-                    const ret = inj.fn.apply(suite);
-
-                    const keys = Object.keys(ret);
-                    const vals = keys.map(function (k) {
-                      return ret[k];
+                  if (err) {
+                    console.error(err.stack || err);
+                    gracefulExit(err);
+                  }
+                  else {
+                    process.nextTick(function () {
+                      suite.__invokeChildren(null, start); //pass start function all the way through program until last child delay call is invoked!
                     });
-
-                    Promise.all(vals).then(function (vals) {
-
-                      keys.forEach(function (k, index) {
-                        suite.injectedValues[k] = vals[index];
-                      });
-
-                      process.nextTick(cb);
-
-                    }, cb);
-
-                  }, function (err) {
-
-                    if (err) {
-                      console.error(err.stack || err);
-                      gracefulExit(err);
-                    }
-                    else {
-                      process.nextTick(function () {
-                        suite.__invokeChildren(null, start); //pass start function all the way through program until last child delay call is invoked!
-                      });
-                    }
-
-                  });
+                  }
 
                 });
-
 
               }
             });
@@ -347,22 +321,6 @@ module.exports = {
       }
 
       function start() {
-
-        // const timeoutVal = global.weAreDebugging ? 50000000 : (global.sumanConfig.defaultTestSuiteTimeout || 100000);
-
-        // const timeoutVal = 50000000;
-        //
-        // const to = setTimeout(function () {
-        //   //note: this timer is necessary so that a test suite will timeout and allow blocked processes to run
-        //   console.error('\n\n', colors.red(' => Suman fatal error => Suite timed out after ' + timeoutVal + ' seconds.'));
-        //   process.exit(constants.EXIT_CODES.SUITE_TIMEOUT);
-        // }, timeoutVal);
-
-        // var count = 0;
-        //
-        // ee.on('test_complete', function handleCompleteTest() {
-        //     count++;
-        // });
 
         function runSuite(suite, cb) {
