+++ lib/helpers/log-exit.js
