+++ lib/helpers/slack-integration.js
