+++ scripts/read-slack-channel.js
