+++ suman.conf.js
