+++ test/testsrc/es5-es6/b.js
