+++ test/testsrc/es5-es6/dep-injection/injected-values/one.test.js
@@ -3,7 +3,6 @@ const Test = suman.init(module, {});
 
 Test.create('example', function (before, describe, inject) {
 
-
     inject(t => {
 
         return {
